Began creating Player/Enemy classes

+++ Game.h
@@ -3,5 +3,5 @@ class Game
 {
 public:
 	bool inBattle;
-	int enemiesDefeated;
+	short enemiesDefeated;
 };
