+++ Player.h
@@ -7,10 +7,10 @@ using namespace std;
 class Player
 {
 public:
-	int health;
-	int magic;
+	short health;
+	short magic;
 	string currentAction;
-	void takeDamage(int damage); //reduces health by the value passed through "damage"
-	int getAttackDamage(); //generates a random integer between 80 and 120
+	void takeDamage(short damage); //reduces health by the value passed through "damage"
+	short getAttackDamage(); //generates a random integer between 80 and 120
 	Player();
 };
